+++ lang/syn/src/codegen/accounts/try_accounts.rs
@@ -400,9 +400,12 @@ fn generate_duplicate_mutable_checks(accs: &AccountsStruct) -> proc_macro2::Toke
             quote! {
                 for key in #composite_name.duplicate_mutable_account_keys() {
                     if !__mutable_accounts.insert(key) {
-                        return Err(anchor_lang::error::Error::from(
-                            anchor_lang::error::ErrorCode::ConstraintDuplicateMutableAccount
-                        ).with_account_name(format!("{}", key)));
+                        return Err(anchor_lang::error::__anchor_with_account_name!(
+                            anchor_lang::error::Error::from(
+                                anchor_lang::error::ErrorCode::ConstraintDuplicateMutableAccount
+                            ),
+                            format!("{}", key)
+                        ));
                     }
                 }
             }
@@ -419,9 +422,12 @@ fn generate_duplicate_mutable_checks(accs: &AccountsStruct) -> proc_macro2::Toke
                 if let Some(key) = #field_keys {
                     // Check for duplicates and insert the key and account name
                     if !__mutable_accounts.insert(key) {
-                        return Err(anchor_lang::error::Error::from(
-                            anchor_lang::error::ErrorCode::ConstraintDuplicateMutableAccount
-                        ).with_account_name(#field_name_strs));
+                        return Err(anchor_lang::error::__anchor_with_account_name!(
+                            anchor_lang::error::Error::from(
+                                anchor_lang::error::ErrorCode::ConstraintDuplicateMutableAccount
+                            ),
+                            #field_name_strs
+                        ));
                     }
                 }
             )*